+++ src/config.js
@@ -1,5 +1,6 @@
 'use strict'
 
+const { MemoryDatastore } = require('interface-datastore')
 const mergeOptions = require('merge-options')
 const Constants = require('./constants')
 
@@ -16,6 +17,9 @@ const DefaultConfig = {
     maxParallelDials: Constants.MAX_PARALLEL_DIALS,
     maxDialsPerPeer: Constants.MAX_PER_PEER_DIALS,
     dialTimeout: Constants.DIAL_TIMEOUT
+  },
+  keychain: {
+    datastore: new MemoryDatastore()
   },
   metrics: {
     enabled: false
